Extract training session
